Fixing behat config
